requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
crispy-bootstrap5==0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.44.0
django-celery-beat==2.2.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
gunicorn==20.1.0
humanize==3.7.0
idna==2.10
kombu==4.6.11
Markdown==3.3.4
oauthlib==3.1.0
pika==1.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
